+++ examples/3qubit_gates.jl
@@ -18,6 +18,7 @@ function test_RX_on_q3()
     "decomposition_type" => "exact",
     
     "optimizer" => "cplex",
+    "optimizer_presolve" => false,
     
     )