+++ CHANGELOG.md
@@ -4,15 +4,15 @@ QuantumCircuitOpt.jl Change Log
 ### v0.1.4
 - Optimizer time limit option added for early termination of solvers
 - `log.jl` updates to support optimizer hitting time limits 
-- Updated commuting matrix valid inequalites 
-- Added support for eliminating pairs of M*M^-1 = M^-1*M = I
+- Revamping of commuting gate valid constraints (dropped support for triplets)
+- Revamping data.jl to fix default values of user inputs
 - Updated unit tests for the above changes
 
 ### v0.1.3
 - Update to minimize_depth from maximizing IGate to minimizing non-IGates, which also improved run times. 
 - Clean-up and update to unit tests to reflect the above change
 - Fixed Gurobi solver bug in `examples/solver.jl`
-- Involutory matrix constraints added and corresponding unit tests
+- Involutory gate constraints added and corresponding unit tests
 
 ### v0.1.2
 - Update to warm start with an initial circuit (without U3 and R gates)